Move react-redux from peerDependencies to optionalPeerDependencies
