+++ packages/react/tsconfig.json
@@ -8,6 +8,7 @@
     "node_modules"
   ],
   "files": [
-    "./src/index.ts"
+    "./src/index.ts",
+    "./src/redux/index.ts"
   ]
 }
